Cleanup withSpaceAndSizing filters
