+++ packages/block-editor/src/filters/with-space-and-sizing/with-space-and-sizing-controls-advanced.js
@@ -1,0 +1,25 @@
+import { createHigherOrderComponent } from "@wordpress/compose";
+import { Fragment } from "@wordpress/element";
+import ControlsAdvanced from "./controls-advanced";
+import { useSupports } from "../../hooks";
+
+const withSpaceAndSizingControlsAdvanced = createHigherOrderComponent( OriginalComponent => {
+
+  return ( props ) => {
+
+    const supports = useSupports( props.name );
+
+    if ( ! supports?.novaBlocks?.spaceAndSizing ) {
+      return <OriginalComponent { ...props } />
+    }
+
+    return (
+      <Fragment>
+        <OriginalComponent { ...props } />
+        <ControlsAdvanced { ...props } />
+      </Fragment>
+    );
+  };
+});
+
+export default withSpaceAndSizingControlsAdvanced;
