+++ packages/block-editor/src/filters/with-space-and-sizing/with-space-and-sizing-custom-props.js
@@ -1,0 +1,23 @@
+import { getSupports } from "../../utils";
+import { getSpacingCSSProps } from "@novablocks/utils";
+
+const withSpaceAndSizingCustomProps = ( element, blockType, attributes ) => {
+
+  const supports = getSupports( blockType.name );
+
+  if ( ! element || ! supports?.novaBlocks?.spaceAndSizing ) {
+    return element;
+  }
+
+  return Object.assign( {}, element, {
+    props: {
+      ...element.props,
+      style: {
+        ...element.props?.style,
+        ...getSpacingCSSProps( attributes )
+      },
+    }
+  } );
+};
+
+export default withSpaceAndSizingCustomProps;
